two app paths moved to core paths
