+++ clients/extension/src/pages/import/index.tsx
@@ -1,6 +1,6 @@
-import { appPaths } from '@clients/extension/src/navigation'
 import ImportQRPage from '@clients/extension/src/pages/popup/pages/import-qr'
 import { AppProviders } from '@clients/extension/src/providers/AppProviders'
+import { corePaths } from '@core/ui/navigation'
 import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 import { createHashRouter, Navigate, RouterProvider } from 'react-router-dom'
@@ -8,16 +8,16 @@ import { createHashRouter, Navigate, RouterProvider } from 'react-router-dom'
 const router = createHashRouter(
   [
     {
-      path: appPaths.root,
+      path: corePaths.root,
       element: <ImportQRPage />,
     },
     {
       path: '*',
-      element: <Navigate to={appPaths.root} replace />,
+      element: <Navigate to={corePaths.root} replace />,
     },
   ],
   {
-    basename: appPaths.root,
+    basename: corePaths.root,
   }
 )
