+++ clients/extension/src/pages/popup/pages/delete-vault/index.tsx
@@ -1,6 +1,7 @@
 import { Button } from '@clients/extension/src/components/button'
 import { useAppNavigate } from '@clients/extension/src/navigation/hooks/useAppNavigate'
 import { useDeleteVaultMutation } from '@clients/extension/src/vault/mutations/useDeleteVaultMutation'
+import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
 import { getVaultId } from '@core/ui/vault/Vault'
 import { ChevronLeftIcon } from '@lib/ui/icons/ChevronLeftIcon'
@@ -17,13 +18,14 @@ const Component = () => {
   const { t } = useTranslation()
   const { colors } = useTheme()
   const navigate = useAppNavigate()
+  const coreNavigate = useCoreNavigate()
   const currentVault = useCurrentVault()
   const deleteVault = useDeleteVaultMutation()
 
   const handleSubmit = () => {
     if (!deleteVault.isPending) {
       deleteVault.mutateAsync(getVaultId(currentVault)).then(() => {
-        navigate('root')
+        coreNavigate('root')
       })
     }
   }
