+++ clients/extension/src/pages/popup/pages/new-vault/index.tsx
