+++ clients/extension/src/vault/components/ActiveVaultGuard.tsx
@@ -1,3 +1,4 @@
+import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import { CurrentVaultProvider } from '@core/ui/vault/state/currentVault'
 import { useCurrentVaultId } from '@core/ui/vault/state/currentVaultId'
 import { useVaults } from '@core/ui/vault/state/vaults'
@@ -5,20 +6,18 @@ import { getVaultId } from '@core/ui/vault/Vault'
 import { ChildrenProp } from '@lib/ui/props'
 import { FC, useEffect } from 'react'
 
-import { useAppNavigate } from '../../navigation/hooks/useAppNavigate'
-
 export const ActiveVaultGuard: FC<ChildrenProp> = ({ children }) => {
   const currentVaultId = useCurrentVaultId()
   const vaults = useVaults()
   const vault = vaults.find(vault => getVaultId(vault) === currentVaultId)
 
-  const navigate = useAppNavigate()
+  const navigate = useCoreNavigate()
 
   const isDisabled = !vault
 
   useEffect(() => {
     if (isDisabled) {
-      navigate('landing')
+      navigate('newVault')
     }
   }, [isDisabled, navigate])
