+++ core/ui/navigation/index.ts
@@ -9,6 +9,7 @@ import { withoutUndefinedFields } from '@lib/utils/record/withoutUndefinedFields
 import { VaultSecurityType } from '../vault/VaultSecurityType'
 
 export const corePaths = {
+  root: '/',
   vault: '/vault',
   joinKeygen: '/join-keygen',
   setupFastVault: '/vault/setup/fast',
@@ -16,6 +17,7 @@ export const corePaths = {
   setupVault: '/vault/setup',
   importVault: '/vault/import',
   keysign: '/vault/keysign',
+  newVault: '/vault/new',
   reshareVault: '/vault/reshare',
   reshareVaultFast: '/vault/reshare/fast',
   reshareVaultSecure: '/vault/reshare/secure',
